+++ src/lib/progress_bar.h
@@ -3,6 +3,6 @@
 
 /* Pass x = -1, y = -1 if you don't plan on repositioning cursor
  * between calls, for more performance. */
-void progress_bar(int x, int y, int width, size_t cur, size_t end);
+void progress_bar(int x, int y, int width, unsigned long cur, unsigned long end);
 
 #endif
